Auth: don't use DB for auth data
If the mechanism used for the rest of the app's data changes, I still
want auth data to be stored in localStorage.

This renames the `auth0.js` file to `auth.js`, and adds handling for
saving and retrieving user data. These new methods are then used
throughout the app.

The `Home` component had to become a class-based component instead of a
function component so that signing out would rerender correctly.